# src/swagger.json
{
	"components": {
...
		"schemas": {
			"CurseResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"lang": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"content",
					"lang"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostCurseRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"errorUUID": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutCurseRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EpisodeResponse": {
				"properties": {
					"episode_id": {
						"type": "number",
						"format": "double"
					},
					"season": {
						"type": "number",
						"format": "double"
					},
					"episode": {
						"type": "number",
						"format": "double"
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"episode_id",
					"season",
					"episode",
					"netflix_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostEpisodeRequest": {
				"properties": {
					"episode": {
						"type": "number",
						"format": "double"
					},
					"season": {
						"type": "number",
						"format": "double"
					},
					"netflixId": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"episode",
					"season",
					"netflixId",
					"movieId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutEpisodeRequest": {
				"properties": {
					"episode": {
						"type": "number",
						"format": "double"
					},
					"season": {
						"type": "number",
						"format": "double"
					},
					"netflixId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"episode",
					"season",
					"netflixId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MovieNameResponse": {
				"properties": {
					"movie_name_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"movie_name_id",
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieResponse": {
				"properties": {
					"movie_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"movie",
							"serie"
						]
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					},
					"minutes": {
						"type": "number",
						"format": "double"
					},
					"movie_names": {
						"items": {
							"$ref": "#/components/schemas/MovieNameResponse"
						},
						"type": "array"
					},
					"episodes": {
						"items": {
							"$ref": "#/components/schemas/EpisodeResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"movie_id",
					"lang",
					"type",
					"minutes",
					"movie_names",
					"episodes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LyricSentenceResponse": {
				"properties": {
					"sentence_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"sentence_id",
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LyricResponse": {
				"properties": {
					"lyric_id": {
						"type": "number",
						"format": "double"
					},
					"movie": {
						"$ref": "#/components/schemas/MovieResponse"
					},
					"episode": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EpisodeResponse"
							}
						],
						"nullable": true
					},
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"sentences": {
						"items": {
							"$ref": "#/components/schemas/LyricSentenceResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"lyric_id",
					"movie",
					"episode",
					"seconds",
					"quality",
					"sentences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostLyricRequest": {
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"episodeId": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"seconds",
					"movieId",
					"quality",
					"episodeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutLyricRequest": {
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"episodeId": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"seconds",
					"movieId",
					"quality",
					"episodeId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"scopes": {
					"contributor": "It allows you to modify (Create, Update, Delete) the content of the application. E.g. Movies, Languages etc.",
					"client": "It allows you to access application data. This permission should be the only one for the end user.",
					"admin": "It allows you to add new keys and permissions. It also allows you to assign permissions keys."
				}
			}
		}
	},
	"info": {
		"title": "lyricated-api",
		"version": "v2 javascript",
		"description": "API for Lyricated app.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/curse/add": {
			"post": {
				"operationId": "PostCurse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurseResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Curse"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostCurseRequest"
							}
						}
					}
				}
			}
		},
		"/curse/find": {
			"get": {
				"operationId": "GetCurses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"curses": {
											"items": {
												"$ref": "#/components/schemas/CurseResponse"
											},
											"type": "array"
										}
									},
									"required": [
										"curses"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Curse"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "only_lang",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/curse/{id}": {
			"get": {
				"operationId": "GetCurse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurseResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Curse"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutCurse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurseResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Curse"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutCurseRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCurse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurseResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Curse"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/episode/new": {
			"post": {
				"operationId": "PostEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpisodeResponse"
								}
							}
						}
					},
					"422": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Episode"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostEpisodeRequest"
							}
						}
					}
				}
			}
		},
		"/episode/{id}": {
			"get": {
				"operationId": "GetEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpisodeResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Episode"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpisodeResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Episode"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutEpisodeRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEpisode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EpisodeResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Episode"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/lyric/new": {
			"post": {
				"operationId": "CreateLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"422": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostLyricRequest"
							}
						}
					}
				}
			}
		},
		"/lyric/without-quality": {
			"get": {
				"operationId": "GetLyricWithoutQuality",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/LyricResponse"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": []
			}
		},
		"/lyric/random": {
			"get": {
				"operationId": "GetRandomLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "qualityBetterThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "qualityLowerThan",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "qualityEqual",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lyric/{id}": {
			"get": {
				"operationId": "GetLyricById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutLyricRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"422": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}